Feat: Set up Passport
Set up Passport to allow user to register or login to an account. Use
Express-session to handle sessions. It is a middleware that saves the
session id as a cookie in the client and allows us to access the session
data using that id on the server. This way we keep personal account
information out of the cookie used by the client to verify to our server
they are authenticated and just keep the key to access the data stored
on the server.